Concurrency Programs
Just some programs I developed at uni, these aim to be examples of actual implementations of concurrency problems to show of the difficulties of making such implementations and designs. The only one which does not achieve concurrency is the Smokers Problem implementation, in which I was dazzled by the visions of a beautiful algorithm which actually removes concurrency, even as such, I think it is far more interesting to see how to not do it.